+++ core/cli/root.go
@@ -11,7 +11,7 @@ import (
 )
 
 // Version is the current framework version.
-const Version = "2.5.0"
+const Version = "2.6.0"
 
 var rootCmd = &cobra.Command{
 	Use:   "rapidgo",